Address merge conflicts
